+++ src/main/java/net/brnbrd/delightful/common/item/food/Nutrition.java
@@ -158,4 +158,7 @@ public class Nutrition {
 	public static final FoodProperties CANTALOUPE_GUMMY = (new FoodProperties.Builder())
 			.nutrition(2).saturationMod(0.0F).alwaysEat()
 			.effect(() -> new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 300, 1), 1.0F).build();
+	public static final FoodProperties BAKLAVA_SLICE = (new FoodProperties.Builder())
+			.nutrition(4).saturationMod(0.35F).fast()
+			.effect(() -> new MobEffectInstance(MobEffects.DIG_SPEED, 600, 0, false, false), 1.0F).build();
 }