+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulBlockTagProvider.java
@@ -37,7 +37,8 @@ public class DelightfulBlockTagProvider extends BlockTagsProvider {
 				.add(DelightfulBlocks.BLUEBERRY_PIE.get())
 				.add(DelightfulBlocks.GREEN_APPLE_PIE.get())
 				.add(DelightfulBlocks.SOURCE_BERRY_PIE.get())
-				.add(DelightfulBlocks.GLOOMGOURD_PIE.get());
+				.add(DelightfulBlocks.GLOOMGOURD_PIE.get())
+				.add(DelightfulBlocks.BAKLAVA.get());
 		this.tag(ModTags.WILD_CROPS)
 				.add(DelightfulBlocks.WILD_SALMONBERRIES.get());